restructure Rust backend into geom/, font/, skia/, and py/ modules
Reorganize the crate from a flat layout into domain-layered modules:

- geom/: pure geometry model (Point, Outline, Bounds, encoding) with no
  external crate dependencies
- font/: skrifa-dependent glyph outline extraction
- skia/: skia-safe dependency isolated to one place (build_skia_path,
  remove_overlaps, render_bitmap, shared math utilities)
- transform/: pure Rust curve transforms, now dep-free from Skia
- py/: PyO3 binding layer only (GlyphDataset, GlyphItem, #[pyfunction]
  wrappers); lib.rs reduced to a 5-line #[pymodule] entry point
- dataset/: internal Rust logic unchanged, visibility widened to pub(crate)
  for items consumed by py/
