+++ src/font/mod.rs
@@ -1,0 +1,3 @@
+mod extract;
+
+pub(crate) use extract::extract_glyph_outline;
