+++ src/geom/mod.rs
@@ -1,0 +1,9 @@
+mod bounds;
+mod encoding;
+mod outline;
+mod point;
+
+pub(crate) use bounds::{Bounds, BoundsPen, bounds_from_outline};
+pub(crate) use encoding::ElementType;
+pub(crate) use outline::{Outline, PathElement, Subpath, SubpathBuilder};
+pub(crate) use point::Point;
