+++ src/geom/outline.rs
@@ -1,54 +1,4 @@
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub(crate) struct Point {
-    pub(crate) x: f32,
-    pub(crate) y: f32,
-}
-
-impl Point {
-    pub(crate) fn new(x: f32, y: f32) -> Self {
-        Self { x, y }
-    }
-
-    pub(crate) fn lerp(self, other: Self, t: f32) -> Self {
-        Self::new(
-            self.x + (other.x - self.x) * t,
-            self.y + (other.y - self.y) * t,
-        )
-    }
-
-    pub(crate) fn midpoint(self, other: Self) -> Self {
-        self.lerp(other, 0.5)
-    }
-
-    pub(crate) fn dot(self, other: Self) -> f32 {
-        self.x * other.x + self.y * other.y
-    }
-
-    pub(crate) fn cross(self, other: Self) -> f32 {
-        self.x * other.y - self.y * other.x
-    }
-
-    pub(crate) fn norm(self) -> f32 {
-        if !self.x.is_finite() || !self.y.is_finite() {
-            return f32::INFINITY;
-        }
-        self.dot(self).sqrt()
-    }
-}
-
-impl std::ops::Add for Point {
-    type Output = Self;
-    fn add(self, other: Self) -> Self {
-        Self::new(self.x + other.x, self.y + other.y)
-    }
-}
-
-impl std::ops::Sub for Point {
-    type Output = Self;
-    fn sub(self, other: Self) -> Self {
-        Self::new(self.x - other.x, self.y - other.y)
-    }
-}
+use super::point::Point;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 #[allow(clippy::enum_variant_names)]
@@ -129,20 +79,20 @@ impl Outline {
     }
 }
 
-pub(super) struct SubpathBuilder {
-    pub(super) start: Point,
-    pub(super) elements: Vec<PathElement>,
+pub(crate) struct SubpathBuilder {
+    pub(crate) start: Point,
+    pub(crate) elements: Vec<PathElement>,
 }
 
 impl SubpathBuilder {
-    pub(super) fn new(start: Point) -> Self {
+    pub(crate) fn new(start: Point) -> Self {
         Self {
             start,
             elements: Vec::new(),
         }
     }
 
-    pub(super) fn finish(self, closed: bool) -> Subpath {
+    pub(crate) fn finish(self, closed: bool) -> Subpath {
         Subpath::new(self.start, self.elements, closed)
     }
 }
