+++ src/geom/point.rs
@@ -1,0 +1,51 @@
+#[derive(Clone, Copy, Debug, Default, PartialEq)]
+pub(crate) struct Point {
+    pub(crate) x: f32,
+    pub(crate) y: f32,
+}
+
+impl Point {
+    pub(crate) fn new(x: f32, y: f32) -> Self {
+        Self { x, y }
+    }
+
+    pub(crate) fn lerp(self, other: Self, t: f32) -> Self {
+        Self::new(
+            self.x + (other.x - self.x) * t,
+            self.y + (other.y - self.y) * t,
+        )
+    }
+
+    pub(crate) fn midpoint(self, other: Self) -> Self {
+        self.lerp(other, 0.5)
+    }
+
+    pub(crate) fn dot(self, other: Self) -> f32 {
+        self.x * other.x + self.y * other.y
+    }
+
+    pub(crate) fn cross(self, other: Self) -> f32 {
+        self.x * other.y - self.y * other.x
+    }
+
+    pub(crate) fn norm(self) -> f32 {
+        if !self.x.is_finite() || !self.y.is_finite() {
+            return f32::INFINITY;
+        }
+        self.dot(self).sqrt()
+    }
+}
+
+impl std::ops::Add for Point {
+    type Output = Self;
+    fn add(self, other: Self) -> Self {
+        Self::new(self.x + other.x, self.y + other.y)
+    }
+}
+
+impl std::ops::Sub for Point {
+    type Output = Self;
+    fn sub(self, other: Self) -> Self {
+        Self::new(self.x - other.x, self.y - other.y)
+    }
+}
