--- src/outline/mod.rs
@@ -1,7 +1,0 @@
-mod encoding;
-mod extract;
-mod model;
-
-pub(crate) use encoding::ElementType;
-pub(crate) use extract::extract_glyph_outline;
-pub(crate) use model::{Outline, PathElement, Point, Subpath};
